refactor: remove duplication & reuse otp methods
